requirements.txt
requests
keystoneauth1
neutronclient.v2_0
